Adapting for delayed linking agent
