+++ examples/study.cases/nek5000/Makefile
@@ -4,22 +4,29 @@ KORALICFLAGS=`python3 -m korali.cxx --cflags`
 KORALILIBS=`python3 -m korali.cxx --libs`
 
 NEKDIR=_deps/nek5000
-NEK5000CFLAGS=
-NEK5000LIBS=$(NEKDIR)/examples/turbChannel/turbChannel.o $(NEKDIR)/examples/turbChannel/libnek5000.a -L$(NEKDIR)/3rd_party/blasLapack -lblasLapack -L$(NEKDIR)/3rd_party/gslib/lib -lgs -Wl,--allow-multiple-definition -lgfortran
+NEK5000LIBS=-shared -L$(NEKDIR)/3rd_party/blasLapack -lblasLapack -L$(NEKDIR)/3rd_party/gslib/lib -lgs -Wl,--allow-multiple-definition -lgfortran
+
+ENVLIBOBJECTS=$(NEKDIR)/examples/turbChannel/obj/*.o $(NEKDIR)/examples/turbChannel/turbChannel.o agent.o
 
 .SECONDARY:
 .PHONY: all 
-all: $(BINARIES)
+all: $(BINARIES) libagent.so
 
-$(BINARIES) : % : %.o environment.o
-	$(KORALICXX) -o $@ $^ $(KORALILIBS) $(NEK5000LIBS)
+$(BINARIES) : % : %.o environment.o 
+	$(KORALICXX) -o $@ $^ $(KORALILIBS) 
+
+agent.o: _environment/agent.cpp
+	$(KORALICXX) -c $< $(KORALICFLAGS) -fPIC
+	
+environment.o: _environment/environment.cpp
+	$(KORALICXX) -c $< $(KORALICFLAGS)
 
 %.o: %.cpp
-	$(KORALICXX) -c $< $(KORALICFLAGS) $(NEK5000CFLAGS)
+	$(KORALICXX) -c $< $(KORALICFLAGS) 
 
-environment.o: _environment/environment.cpp
-	$(KORALICXX) -c $< $(KORALICFLAGS) $(NEK5000CFLAGS)
+libagent.so: $(ENVLIBOBJECTS)
+	$(KORALICXX) -o $@ $^ $(NEK5000LIBS) 
  
 .PHONY: clean
 clean:
-	$(RM) $(BINARIES) *.o *.ti *.optrpt *.h5 *.dat *.xmf
+	$(RM) $(BINARIES) *.o *.so *.ti *.optrpt *.h5 *.dat *.xmf
