+++ examples/study.cases/nek5000/install_deps.sh
@@ -29,8 +29,8 @@ popd
 cp _config/* ${NEK5000_DIR}/examples/turbChannel
 cp _environment/turbChannel.usr ${NEK5000_DIR}/examples/turbChannel
 pushd ${NEK5000_DIR}/examples/turbChannel
-MPI=0 FFLAGS='-O3 -g' CFLAGS='-O3 -g' ${NEK5000_DIR}/bin/nekconfig -build-dep;
-MPI=0 FFLAGS='-O3 -g' CFLAGS='-O3 -g' ${NEK5000_DIR}/bin/nekconfig;
+MPI=0 FFLAGS='-O3 -g -fPIC' CFLAGS='-O3 -g -fPIC' ${NEK5000_DIR}/bin/nekconfig -build-dep;
+MPI=0 FFLAGS='-O3 -g -fPIC' CFLAGS='-O3 -g -fPIC' ${NEK5000_DIR}/bin/nekconfig;
 make -j 6 lib usrfile 
 popd
